Move bpf FD mount to unix-specific file
